# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/provider-skills": {
      "get": {
        "operationId": "ProviderSkillsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "x-role": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "Get all provider skills (scoped for managers)",
        "tags": [
          "provider-skills"
        ]
      },
      "post": {
        "operationId": "ProviderSkillsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProviderSkillDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created successfully"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "x-role": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "Assign skill to provider",
        "tags": [
          "provider-skills"
        ]
      }
    },
    "/provider-skills/provider/{provider_id}": {
      "get": {
        "operationId": "ProviderSkillsController_findByProvider",
        "parameters": [
          {
            "name": "provider_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "x-role": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "Get provider skills by provider ID",
        "tags": [
          "provider-skills"
        ]
      }
    },
    "/provider-skills/skill/{skill_id}": {
      "get": {
        "operationId": "ProviderSkillsController_findBySkill",
        "parameters": [
          {
            "name": "skill_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "x-role": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "Get providers by skill ID",
        "tags": [
          "provider-skills"
        ]
      }
    },
    "/provider-skills/verify/{id}": {
      "patch": {
        "operationId": "ProviderSkillsController_verifySkill",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyProviderSkillDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "x-role": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "Verify provider skill",
        "tags": [
          "provider-skills"
        ]
      }
    },
    "/provider-skills/reject/{id}": {
      "patch": {
        "operationId": "ProviderSkillsController_rejectSkill",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyProviderSkillDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Skill request rejected and removed"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "x-role": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "Reject provider skill verification request",
        "tags": [
          "provider-skills"
        ]
      }
    },
...
    "/job-assignments": {
      "get": {
        "operationId": "JobAssignmentsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "Get all job assignments",
        "tags": [
          "job-assignments"
        ]
      },
      "post": {
        "operationId": "JobAssignmentsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateJobAssignmentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "Create job assignment",
        "tags": [
          "job-assignments"
        ]
      }
    },
    "/job-assignments/provider/{sp_id}": {
      "get": {
        "operationId": "JobAssignmentsController_findByProvider",
        "parameters": [
          {
            "name": "sp_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "Get job assignments by provider",
        "tags": [
          "job-assignments"
        ]
      }
    },
    "/job-assignments/{id}": {
      "get": {
        "operationId": "JobAssignmentsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "Get job assignment by ID",
        "tags": [
          "job-assignments"
        ]
      },
      "patch": {
        "operationId": "JobAssignmentsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateJobAssignmentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "Update job assignment",
        "tags": [
          "job-assignments"
        ]
      },
      "delete": {
        "operationId": "JobAssignmentsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "Delete job assignment",
        "tags": [
          "job-assignments"
        ]
      }
    },
    "/job-assignments/{id}/status": {
      "patch": {
        "operationId": "JobAssignmentsController_updateStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "Update job assignment status",
        "tags": [
          "job-assignments"
        ]
      }
    },
...
  },
  "info": {
    "title": "TatkuUnited API",
    "description": "Service platform API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Paste your JWT token (without the Bearer prefix)"
      },
      "x-role": {
        "type": "apiKey",
        "in": "header",
        "name": "x-role",
        "description": "Role header required by RolesGuard. Allowed values: super_user, region_manager, service_provider, customer"
      }
    },
    "schemas": {
...
      "CreateProviderSkillDto": {
        "type": "object",
        "properties": {
          "service_provider_id": {
            "type": "string",
            "example": "provider-uuid"
          },
          "skill_id": {
            "type": "string",
            "example": "skill-uuid"
          },
          "is_verified": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "service_provider_id",
          "skill_id"
        ]
      },
      "VerifyProviderSkillDto": {
        "type": "object",
        "properties": {
          "skill_id": {
            "type": "string",
            "example": "skill-uuid"
          }
        },
        "required": [
          "skill_id"
        ]
      },
...
      "CreateJobAssignmentDto": {
        "type": "object",
        "properties": {
          "booking_id": {
            "type": "string"
          },
          "service_id": {
            "type": "string"
          },
          "sp_id": {
            "type": "string"
          },
          "scheduled_date": {
            "type": "string"
          },
          "hour_start": {
            "type": "string"
          },
          "hour_end": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          }
        },
        "required": [
          "booking_id",
          "service_id",
          "sp_id",
          "scheduled_date",
          "hour_start",
          "hour_end"
        ]
      },
      "UpdateJobAssignmentDto": {
        "type": "object",
        "properties": {
          "booking_id": {
            "type": "string"
          },
          "service_id": {
            "type": "string"
          },
          "sp_id": {
            "type": "string"
          },
          "scheduled_date": {
            "type": "string"
          },
          "hour_start": {
            "type": "string"
          },
          "hour_end": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          }
        }
      },
...
    }
  }
}